+++ CHANGELOG.md
@@ -53,6 +53,7 @@ Line wrap the file at 100 chars.                                              Th
   Previously, the installer would abort.
 - Revert to using netsh for DNS config, as some Windows builds did not deal with changes correctly.
   `TALPID_DNS_MODULE` can be used to override this.
+- Fix deadlock when updating split tunnel addresses.
 
 ### Changed
 - Update Electron from 19.0.13 to 21.1.1.